bugdown: Move links related function to links.py.
